make a distinction on what is supported
When i read the table initially, it wasn't clear that the supported scenarios are applicable when the UPN differs in AD/AAD. I was looking to move from ADFS (Federated) to PHS/SSSO (Managed) and I initially thought this scenario would be unsupported. Making this distinction in the table description will clear up any confusion.